requirements.txt
streamlit==1.32.0
torch
torchvision
pillow
numpy

